+++ lib/fields/hours/hours_page.dart
@@ -94,7 +94,9 @@ class _OpeningHoursPageState extends ConsumerState<OpeningHoursPage> {
     return Padding(
       padding: EdgeInsets.all(10.0),
       child: TextFormField(
-        initialValue: hours.hours,
+        // The "24/7" value is removed in `initState()`, so we return it here.
+        // (Although it shouldn't be called logically).
+        initialValue: widget.hours?.trim() == '24/7' ? '24/7' : hours.hours,
         textCapitalization: TextCapitalization.words,
         keyboardType: TextInputType.visiblePassword,
         autovalidateMode: AutovalidateMode.always,