+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 _Unreleased_
 
-### Opening Hours Editor
-
 * Could not mark an hours fragment as `off`.
 * When switching to the raw hours editor, the value was not updated.
 * An inactive `PH` interval is no longer added by default.
+* `24/7` value was missing from the raw value editor.
+* Removed `addr:postcode` and added `building:material` to the building properties.
+* Removed non-card options from the payment setting.
+* Disabled map jumping in POI mode because of the number of POI.
 
 ## 0.7